Unify `Relation#klass` and `Relation#model`
One is the alias of the other, and have been so for a very long time.

My issue with this is that `klass` is really a bad name, and it's
very confusing when reading code to see a mix of both names being
used.

There is no point officially deprecating either, but inside
Active Record we should consistently only use one, I chose to
use `model` as it's way more descriptive in my opinion.