Switch example editor to code instead of mate
It's been almost 10 years since the last TextMate update. 😢

I've seen developers be confused here - "huh, what's mate?" - and while I do not want to be seen as advocating for a particular editor, I would bet that more developers are on VS Code than are still holding on to TextMate.