Use the env variable in the key name for view cache
This is to prevent any cross contamination between the environments with regards to the cache keys.  An app  using the same redis instance for the cache between two environments and values for one were making it into the values for another.  Instead of remembering to use a different instance/db or emptying the instance, a better way would just be to specify the environment in the cache key.

update changelog file