Avoid alias `exec_update` and `exec_delete`
During a recent incident, I was trying to clean up some data in a production console via an `update_all`. The query raised (due to a lock timeout), and I noticed something very surprising in the backtrace - a call to `exec_delete`!

It turns out that these methods are aliases becuase they both have a very simple and identical implementation. But I didn't know this at a time and my heart definitely skipped a beat.

I am proposing we avoid aliasing for the sake of avoid confusion in backtraces. If you call `update`, I think we should avoid showing any other SQL verbs in your backtrace if we can.

(I realise this is a super nitpicky change so I won't be upset if it gets rejected/ignored.)