Add mariadb options to application generator
Until now, the database option has conflated the datbase adapter and the database itself. It's fine for postgres and sqlite, but for mysql2 and trilogy adapters we could either use mysql or mariadb as the database. This becomes a problem for generating a dev container, since we need to add a service to docker compose with the correct database. Previously, we gave the user a mysql database for the mysql option, and a mariadb database for the trilogy option. But that is no good as it is common to use a mysql database with trilogy.

In this PR I disambiguate the adapter-database issue by introducing two new options, mariadb-mysql and mariadb-trilogy. When generating a dev container, the mysql and trilogy options will create a mysql database, and the maria specific options will generate a mariadb database. Users not generating a dev container do not need to worry about the new options and can continue using the original options, as this change only effects dev container setup.