Reallow setting secret_key_base to nil when local
Previously, secret_key_base was allowed to be set to `nil` in local
environments (or with SECRET_KEY_BASE_DUMMY) because validation would
only happen on usage and not on the setter. This was recently
[changed][1] to make it easier to identify exactly where a
secret_key_base was being set to an invalid value.

However, this broke some applications which unconditionally set
secret_key_base to some external value in dev/test. Before the change,
the set value could be `nil` and fall back to the generated local secret
on usage.

This commit restores that behavior so that applications can continue to
set secret_key_base unconditionally, since the nil value will end up
getting replaced by the generated local secret anyways.

[1]: c2901eb084adb3d3701be157bec20d2961beb515