FIX: Sessions generator opening new and create page
If current session is present, redirect to root path on: new or create.
We can't rely on `Current.session` since authentication is skiped above,
we need to check if user has a session cookie already.

I'm not sure whether using `resume_session` is the best approach here but
it was the one with the least amount of ceremony.