Always discard Trilogy multi-statement results
Previously we were trying to clear the extra results in a
with_raw_connection block, which caused some issues.

We never want to connect or reconnect here, we just need to drop any
results from a multi-statement query and that only makes sense to do on
the same connection we'd already used for the query.

Previously when a query was run which hit the query cache on a fresh,
non-sticky connection, the connection would perform its `verfiy!` before
calling next_result, resulting in a network round trip.