Reflow Active Storage transforms section [ci-skip]
Both "optional arguments" and "migrating processors" seem too specific
for where they were previously located within the section. Both ideas
are more complicated/processor specific, but are located very early on
when the ideas should be more simple/introductory.

This commit reorders the paragraphs a bit to improve the section's flow.
Now, the most basic examples are presented first, then the idea of
processors and what they have in common, and finally the more advanced
options and how the processors differ.