Refactor `PostgresqlAdapter#raw_execute` to be less unique
The postgres adapter used to be more complex than the others to
be able to pass the prepared statement key to the `log` method.

If we instead add it to the payload later, we can simplify the method
further and it opens the door to refactor `log` and `with_raw_connection`
out of `raw_execute`.