Keep using log in DatabaseStatements#raw_execute
The payload construction was moved here in a previous commit, but log
wasn't removed or deprecated.

We've been using this as a hook we can override to add some adapter
specific-keys. If we want to remove or bypass this at a later date we
should probably introduce something that can be overridden like we did
for cache_notification_info.

Co-authored-by: Daniel Colson <danieljamescolson@gmail.com>