Refactor SQLite3Adapter to simplify transaction management
Followup: https://github.com/rails/rails/pull/52428

This adapter was overly complex following https://github.com/rails/rails/pull/37798.

I'm also not convinced storing the `read_uncommited` value in the
execution state was really valid, it's a connection property.

We can assume outside of transaction, `read_uncommited` is set to
the default, and simply reset it at the end of any transaction
that did change it.

This makes SQLite3Adapter much less of a snowflake.