Get rid of `with_multi_statement`
It's a noop on all adapters except MySQL which doesn't allow multi
statements by default.

Also `batch_execute` in Sqlite3Adapter has to use many ow level APIs
because while sqlite3 allow multi statement by default, you have to use
a different API for that.

So overall `with_multi_statement` was always a lie, because it wouldn't
work on sqlite3 and it's simpler to pass a `batch: true` argument to `raw_execute`
so that each adapter can do the necessary work to allow multi-statement.

This will simplify a future PR that Adrianna will open soon.

Co-Authored-By: Adrianna Chang <adrianna.chang@shopify.com>