Improve documentation of `ActionView:RenderingHelper#render`
The current documentation `RenderingHelper#render` doesn't document all
options and documents an the outdated `:update` argument.

The documentation of `ActionController::Rendering#render` is much more
complete. It shows example code for the different rendering modes, has a
separate section for other arguments and reads a lot better as well.

This change copies the documentation from `ActionController::Rendering`
and updates it to make sense for ActionView:
- The `response formats` is replaced with `formats`.
- The `assigns`, `status` and `layout` options are removed as the aren't
  supported.
- The `:formats`, `:variants`, and `:handlers` options are added.

This also removes the `:update` option which is no longer supported.