Preserve isolation level when retrying MySQL transactions
Fix: https://github.com/rails/rails/pull/52501

Since MySQL require to set the isolation level as a statement
before the `BEGIN`, if the `begin` statement fails and we retry it
we're opening a transaction with the default level.

To handle this we can send both statement at once so that if we
retry both are sent.

A much simpler alternative would be to just disallow retrying `BEGIN`
when opening an isolated DB transaction.