Eagerly prune the query caches store
Fix: https://github.com/rails/rails/issues/52617

It isn't per say a leak because the connection reaper eventually prune
these, and also it's not expected that new Threads are constantly
making entries into that cache. But it's true that with Fiber it's
probably a bit more common, and the default reaper frequency isn't
adapter to clear this fast enough.

So we should instead eagerly prune it, or use a WeakMap.