Get rid of Active Record connection reaper
The reaper was introduced back in cde7692d4e3e0e67e480cc6172f6e2bacaceeb5e
The goal is to detect checked out connections owned by dead threads, and
make them available again.

To do this, a thread check connections every 60 seconds, which IMO isn't
nearly enough. But more importabtly since https://github.com/rails/rails/pull/14360
the reaper isn't really useful because in the checkout path, if we can't
immediately checkout a connection, we execute the reaping routine, so IMO
it's not really justified to have a background thread for this.