Move "avoid N+1 queries" explanation lower in the Active Storage documentation
The explanations about avoinding N+1 queries seems out of place early in
the documentation. The actual association hasn't been explained yet and
yet we explain performance optimizations.
It's also confusing when reading the following paragraph where "this
relationship" might seem to refer to the `with_attached_avatar` scope.

Instead we can group it as a performance section together with the
`strict_loading` section.