Remove `private` call
Below this `private` call we only have a class definition, which as we know it
is not affected (`private` only affects methods).

We discussed with @dhh the original intention.

`ActionCable::RemoteConnections::RemoteConnection` is public interface, it is
documented in the API and it is the class of objects returned by the `where`
method. Its documentation explains which is its purpose and restricted public
interface, so we keep it public and only remove the `private` call.