Dispatch direct-upload events on attachment uploads
When using Action Text's rich textarea,  it's possible to attach files to the
editor. Previously, that action didn't dispatch any events, which made it hard
to react to the file uploads. For instance, if an upload failed, there was no
way to notify the user about it, or remove the attachment from the editor.

This commits adds new events - `direct-upload:start`, `direct-upload:progress`,
and `direct-upload:end` - similar to how Active Storage's direct uploads work.

Closes #37793
Supersedes #37794

Co-authored-by: Brad Rees <github@bradleyrees.com>