Add devcontainer to hidden_namespaces
The devcontainer generator is primarily intended to be used with the devcontainer command, which
infers the correct options based on the application configuration. So, let's hide the devcontainer
namespace from the generate command. This doesn't actually prevent the generator from being used
this way, but we don't want to advertise it, since the command will lead to a better experience
for the typical user.

Co-authored-by: Jon Roberts <jon.roberts.8@gmail.com>