Make sure we are using the correct connection pool when dumping the schema
This was broken by 175bf8259a83853836657c0fbf300dfa3842e535.

Before we were switching the connection pool to use the one for the
correspondent db config, but since we moved the `dump_schema` call to
outside of the `with_temporary_pool` block, we were using the default
connection pool.

Fixes #52776.