Do more request parameter parsing ourselves
We still rely on Rack to actually read the input stream, but by
constructing the parameter hashes ourselves, we restore our ability to
extend the meaning of parameter-name formatting in the future, while
also concentrating all our preprocessing into a single walk.

This is an important step in defining a more consistent boundary to our
Rack dependency, allowing them to evolve their behaviour with less
chance of creating some edge-case incompatibility.