Refactor implementation of attributes_for_inspect default values
This is a follow up to rails/rails@e50182a. Based on feedback we should not use consider_requests_local for this. Instead we will set it to :all by default and :id in production (in production.rb). This will also allow this feature to be opt-in for upgrading apps.

Additionally, we will always set it to :all when running the console in production, to avoid interfering with debugging.

Co-authored-by: Levi Cole <lckennedy@gmail.com>