Use implication for skip_active_record/skip_solid
When skip_solid was added, it used the "old style" option implication
where skip_*? methods delegate to all the options that should be
implied. This has a few downsides such as making it unclear to users
when conflicting options may be specified (for example,
skip_active_record and no_skip_solid).

This commit changes the skip_solid option to use the new style of
implication so that error messages are better and the predicate method
is simpler.

Co-authored-by: Jon Rowe <hello@jonrowe.co.uk>