Avoid repeated dup of as_json options
Previously for every element in an Array and every value in a Hash we
would duplicate the options hash passed in.

I did this very conservatively, duping and freezing the Hash when it is
not already frozen so that any cases mutating the Hash should be caught.
I don't feel very strongly about this and we could remove this in the
future.