Deprecate `Benchmark.ms` and add `benchmark` to the gemspec
Ruby plans to make `benchmark` a bundled gem: https://github.com/ruby/ruby/pull/11492

It doesn't make much sense to continue this core extension since rails would
need to add it to its gemspec to continue offering it for what basically amounts to `x * 1000`.

Since requiring it could emit a warning, add it to the gemspec until it can later be removed
again with Rails 8.1

The benchmark generator will continue to work as usual:
It adds `benchmark-ips` to the users Gemfile.