Wrap database seeding in a reload
It is quite possible that database seeding may run jobs inline. In that case a reload will be attempted. For app that is set to reload_classes_only_on_change this would normally not initiate a reload, since no files would have been changed. However, it is common to run the seeds as part of one command that also sets up the database (eg bin/rails db:prepare db:seed). In that case, the schema.rb file may have changes. By default that file is watched by the reloader. So, the application will be reloaded at the time the first job (or other reload) occurs. Reloading the application during the middle of running seeds can lead to unexpected and buggy behaviour.

Instead, we should wrap the entire seeds task in a reload so that the reload only occurs at the end of seeding.