Fix issues with scrolling behaviour
To "restore" the current behaviour on main, a solution was needed
for browser-OS combinations that do not allow for two concurrent
scrolling elements. Current code on main breaks completely when
using scroll-behaviour: smooth (in CSS), and breaks when scrolling
back-to-top from the bottom of the page (highlight no longer
correct).

This commit works around the limitation whilst also enabling the
non-smooth scrolled solution for those who want reduced motion.
Additionally it has a workaround for a bug on Android, and some
defaults for those without JavaScript enabled.