Do not include redis by default in dev container
Rails 8 will use the solid gems by default for job queue and action cable, which do not depend on redis. So we won't include redis in the devcontainer by default. Redis will be included in the app is generated with the skip-solid option and uses active job or action cable. For existing apps, we won't use redis if either of the solid gems are in use.