Loosen time assertions
The time it takes to come back from sleep can be quite variable, but we
can guarantee that it is _at least_ that long.

I loosened this from a 0.5 millisecond buffer to a 10 millisecond
buffer, it's possible that isn't even enough on some platforms.

I also shortened the sleep duration from 100ms each to 10ms, to make our
test suite faster.