Exclude _process_options options in render_json
When :callback, :content_type, or :status are also passed when caling
render :json, those keys are used to control aspects of the response and
aren't intended to be used for the JSON generation itself.

Removing these options allows us to hit the fast path in JSON generation
where we can avoid an extra deep-dup through calling .as_json.